Fix goroutine leak
This PR modifies the `Inform()` function to take in a [Context](https://pkg.go.dev/context) and exit when a signal comes in on the `ctx.Done()` channel. This fixes an issue where the goroutine is not cleaned up when a container restarts or is redeployed. Without this change, additional goroutines are spawned whenever a container restarts, and the previously spawned goroutine still remains active.